Add Vara Chat. Disable wine window management.
Fixed some graphical errors with Vara Chat by disabling Wine window management.  Speed wasn't great in the past (with dotnet) with Wine window mangement disabled, but wine-mono is efficient enough that we can disable it.

Also made some some small edits and added some groundwork for pi3 to be added in the future.