allow local download if megadownload fails
the mega hosting isnt a fan of scripts. this is a patch to allow hunting the downloads folder, before mega-download fails and script crashes out. I didnt add any comments or details to the header or readme about this, minimal change suggestion. the || error didnt work anyway no trap'n ideally documented as well for end user